+++ src/components/Dropdown/DropdownConstants.js
@@ -1,1 +1,3 @@
 export const defaultCustomStyles = baseStyles => baseStyles;
+
+export const ADD_AUTO_HEIGHT_COMPONENTS = ["container", "control", "valueContainer"];