+++ src/components/Link/__tests__/link-snapshot-tests.jest.js
@@ -3,43 +3,38 @@ import renderer from "react-test-renderer";
 import Link from "../Link";
 
 describe("Link renders correctly", () => {
-  it("with empty props", () => {
-    const tree = renderer.create(<Link />).toJSON();
+  it("with icon", () => {
+    const tree = renderer.create(<Link icon="fa fa-star" />).toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+
+  it("with right icon", () => {
+    const tree = renderer.create(<Link icon="fa fa-star" iconPosition={Link.position.END} />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it("with id", () => {
-    const tree = renderer.create(<Link id="id" />).toJSON();
+    const tree = renderer.create(<Link id="TestId" />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
-  it("with className", () => {
-    const tree = renderer.create(<Link componentClassName="my-classname" />).toJSON();
+  it("with componentClassName", () => {
+    const tree = renderer.create(<Link componentClassName="testClassName" />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
-  it("with href", () => {
-    const tree = renderer.create(<Link href="http://somelink.com" />).toJSON();
+  it("with text", () => {
+    const tree = renderer.create(<Link text="Link" />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
-  it("with rel", () => {
-    const tree = renderer.create(<Link rel="something" />).toJSON();
+  it("with ariaLabeledBy", () => {
+    const tree = renderer.create(<Link ariaLabeledBy="aria label link" />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
-  it("with onClick", () => {
-    const tree = renderer.create(<Link onClick={() => {}} />).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-
-  it("with target", () => {
-    const tree = renderer.create(<Link target={Link.target.TOP} />).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-
-  it("a11y", () => {
-    const tree = renderer.create(<Link ariaLabelDescription="My Link" ariaLabeledBy="anotherId" />).toJSON();
+  it("with ariaLabelDescription", () => {
+    const tree = renderer.create(<Link ariaLabelDescription="arialabel link" />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 });